Update extension to version 0.0.15
# Set default dimensions for inserted images (1942135)

# Add missing html field name: 'tip' (78a3a9d)
It would be better to use semantics and check if the `widget` is html and
`type` is text, but that is a little awkward.

Images included in tip fields should now save correctly.

# Use proxy to override semantic entry properties (280cc81)
The overrides modify values in-place. They are sent a deep copy of
semantic entries.

This is a simpler and more flexible way to perform property overrides.

Limit the question types for question sets to MultiChoice, Blanks,
and TrueFalse

# Add error messages to value assertions in FileContentStorage (4a32e17)

# Indicate save errors more explicitly (7230142)
Make the save button turn red and have a different icon

# Use labels instead of headers for metadata fields (6495e26)

# Include attachments from collaborator solutions in metadata.json (aaa4db4)