Simplify handling of private fields
This way you can specify a library-specific answer yanker and
a separate set of private keys if necessary

When a library-specific answer yanker exists, the less specific keys yanker
will be chained to it and the private objects will be merged.

This allows the privacy of `AdditionalField` to be based on its definition

Prior to this change, the keys for the metadata fields needed to be specified
separately from the fields themselves.