Update to 0.0.21
- Update test description (1a0d182)

- Make default merge error when one value is not null (0b78eec)

- Add unit tests; narrow ways of defining private fields (2d7626c)
There should be one--and preferably only one--obvious way to do it.
Special cases aren't special enough to break the rules.
Although practicality beats purity.
Errors should never pass silently.
In the face of ambiguity, refuse the temptation to guess.

- Make AnswerYankers.chain variadic and add unit test (9bd66cb)

- Explicitly require an answer yanker to be defined on a library (da4da83)

- Simplify handling of private fields (e1b9db0)
This way you can specify a library-specific answer yanker and
a separate set of private keys if necessary

When a library-specific answer yanker exists, the less specific keys yanker
will be chained to it and the private objects will be merged.

This allows the privacy of `AdditionalField` to be based on its definition

Prior to this change, the keys for the metadata fields needed to be specified
separately from the fields themselves.

- Remove TODO describing finished work (abb1d06)

- Update answer yanker tests (8305b3b)
* Move question set test into different file because it is implementation-specific
* Make yankByKeysFactory test work a bit better

- Make summary/detailed solution private (a83b383)
Remove unused unyankers (replaced by recursive merge)

- Use recursive merge to merge public and private content (0bbc35c)
Should match recursive merge used in Enki

Also fix the QuestionSet yanker to work with recursive merge

- Merge pull request #15 from openstax/change-license (b5a7580)
change license to AGPL
- change license to AGPL (fe85487)